+++ docs/guides/changelog.md
@@ -9,7 +9,7 @@ description: |-
 
 For migration guidance and non-schema changes, see the [project changelog](https://github.com/chalk-ai/terraform-provider-chalk/blob/main/CHANGELOG.md).
 
-## Unreleased
+## v1.0.11
 
 ### Resources